Factor out zernike code into submodule.
