+++ galsim/__init__.py
@@ -172,3 +172,4 @@ from . import cdmodel
 from . import optics
 from . import utilities
 from . import fft
+from . import zernike
